Rename user_settings to settings.
